Add files via upload
In this project, we conducted an analysis of Ireland's agricultural sector and compared it to other countries worldwide. Our analysis included forecasting, sentiment analysis, and evidence-based recommendations for the sector. We also provided a complete rationale of the entire process used to discover our findings. Using a variety of data sources, including agriculture output in Ireland, fertiliser price, agriculture output in Europe, and land sale price, we were able to gain a comprehensive understanding of the current state of the Irish Agri sector. We found that Ireland has a strong agricultural output, with a steady supply of high-quality products. However, we also identified several areas for improvement, including the need for more efficient use of fertilisers and the potential for increased land productivity through the adoption of new technologies. Overall, our analysis provides valuable insights for decision-makers in the Irish Agri sector, as well as for stakeholders interested in the future of agriculture in Ireland. By leveraging the power of big data, machine learning, and artificial intelligence, we were able to generate evidence-based recommendations that could help drive the sector forward and enhance its competitiveness on a global scale.